api/landing_to_stage/app/requirements.txt
# Function dependencies, for example:
# package>=version
pytz==2023.3
python-dotenv==1.0.0
functions-framework==3.4.0
fsspec==2023.6.0
google-cloud-storage==2.10.0
google-cloud-logging==3.5.0
google-cloud-bigquery==3.11.3